Changement du type de l'id

+++ Craft2Gether/src/main/java/fr/crafttogether/repositories/ListeRepository.java
@@ -3,5 +3,5 @@ package fr.crafttogether.repositories;
 import fr.crafttogether.models.Liste;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ListeRepository extends JpaRepository<Liste, Long> {
+public interface ListeRepository extends JpaRepository<Liste, Integer> {
 }
